+++ lib/mineshaft/downloader.rb
@@ -2,7 +2,7 @@
 
 # author:: Cameron Testerman
 # email:: cameronbtesterman@gmail.com
-# created:: 2017-04-14 1:19PM
+# created:: 2025-01-05 7:49PM
 
 # Copyright (c) 2017 Cameron Testerman
 #
@@ -23,14 +23,32 @@
 # ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 # OTHER DEALINGS IN THE SOFTWARE.
 
-require 'mineshaft/constants'
-require 'mineshaft/logger'
-require 'mineshaft/options'
-require 'mineshaft/date'
-require 'mineshaft/downloader'
-require 'mineshaft/environment'
-require 'mineshaft/unzipper'
-require 'mineshaft/installer'
-require 'mineshaft/version'
-require 'mineshaft/ruby_versions'
-require 'mineshaft/commands'
+module Mineshaft
+  module Downloader
+    def self.download(version:, destination:)
+      LOGGER.log "🪄  Downloading Ruby #{version}..."
+
+      site = Mineshaft::RubyVersions.site
+      zipfile = Mineshaft::RubyVersions.zipfile(version)
+
+      download_file(site, zipfile, destination) do |response|
+        write_file(destination, response)
+      end
+
+      LOGGER.log "🎉 Ruby #{version} successfully downloaded!"
+    end
+
+    def self.write_file(destination, response)
+      open("#{destination}/#{Mineshaft::Installer::RUBY_ARCHIVE}", 'w') do |f|
+        f.write(response.body)
+      end
+    end
+
+    def self.download_file(site, file, destination)
+      Net::HTTP.start(site) do |http|
+        response = http.get(file)
+        yield response
+      end
+    end
+  end
+end