+++ spec/date_spec.rb
@@ -1,8 +1,9 @@
-require 'spec_helper'
-require 'mineshaft/date'
+# frozen_string_literal: true
 
-describe Mineshaft::Date do
-  it "should prepend zero to number" do
-    expect(Mineshaft::Date.prepend_zero(1)).to eql("01")
+require 'spec_helper'
+
+RSpec.describe Mineshaft::Date do
+  it 'should prepend zero to number' do
+    expect(Mineshaft::Date.prepend_zero(1)).to eql('01')
   end
 end